Fix MetadataExtractor issue

+++ src/Tizen.Multimedia/Interop/Interop.MetadataExtractor.cs
@@ -27,9 +27,9 @@ internal static partial class Interop
 		internal static extern int GetFrame(IntPtr handle, out IntPtr frame, out int size);
 
 		[DllImport(Libraries.MetadataExtractor, EntryPoint = "metadata_extractor_get_synclyrics")]
-		internal static extern int GetSynclyrics(IntPtr handle, int index, out ulong timeStamp, out string lyrics);
+		internal static extern int GetSynclyrics(IntPtr handle, int index, out uint timeStamp, out string lyrics);
 
 		[DllImport(Libraries.MetadataExtractor, EntryPoint = "metadata_extractor_get_frame_at_time")]
-		internal static extern int GetFrameAtTime(IntPtr handle, ulong timeStamp, bool isAccurate, out IntPtr frame, out int size);
+		internal static extern int GetFrameAtTime(IntPtr handle, uint timeStamp, bool isAccurate, out IntPtr frame, out int size);
 	}
 }
